Use django_rq to replace thread starting a request
