+++ main/approval/tests/functional/test_requst_end_points.py
@@ -1,6 +1,7 @@
 """ Module to test approval requests and actions """
 import pytest
 import json
+import django_rq
 from django.urls import reverse
 from main.approval.services.create_request import CreateRequest
 from main.tests.factories import default_tenant
@@ -88,9 +89,7 @@ def test_request_action_detail(api_request):
 
 @pytest.mark.django_db
 def test_create_request(api_request, mocker):
-    mocker.patch.object(
-        CreateRequest, "_CreateRequest__prepare_request", return_value=None
-    )
+    mocker.patch("django_rq.enqueue")
     default_tenant()
     url = reverse("request-list")
     response = api_request(
